updated Game with Room
